refactor: rename files and remove some comments
