+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "build": "NODE_ENV=production webpack --config webpack.prod.js",
     "dev": "NODE_ENV=development webpack --config webpack.dev.js",
-    "merkle": "ts-node src/server/mock-merkle-proof.ts",
+    "merkle": "ts-node src/server/mockMerkleProof.ts",
     "test": "jest",
     "test:coverage": "jest --coverage",
     "test:watch": "jest --watch",
@@ -83,7 +83,6 @@
     "webpack-merge": "^5.8.0"
   },
   "dependencies": {
-    "@dimensiondev/metamask-extension-provider": "https://github.com/DimensionDev/extension-provider/tarball/master",
     "@ethersproject/address": "^5.7.0",
     "@ethersproject/providers": "^5.7.2",
     "@ethersproject/units": "^5.7.0",
@@ -105,6 +104,7 @@
     "fast-deep-equal": "^3.1.3",
     "link-preview-js": "^3.0.4",
     "loglevel": "^1.8.1",
+    "metamask-extension-provider": "^3.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-jazzicon": "^1.0.4",
