+++ src/background/controllers/browserUtils.ts
