+++ src/background/services/simpleStorage.ts
