+++ src/background/services/zkValidator.ts
