--- src/background/shared/checkManifestV3.ts
@@ -1,3 +1,0 @@
-import { browser } from "webextension-polyfill-ts";
-
-export const isManifestV3 = browser.runtime.getManifest().manifest_version === 3;
