+++ src/background/shared/initGlobals.ts
