+++ src/contentScripts/proof/index.ts
