+++ src/contentScripts/proof/semaphore.ts
