+++ src/setupTests.ts
@@ -1,1 +1,9 @@
 import "@testing-library/jest-dom";
+
+jest.mock("loglevel", () => ({
+  info: jest.fn(),
+  log: jest.fn(),
+  debug: jest.fn(),
+  warn: jest.fn(),
+  error: jest.fn(),
+}));
