+++ src/ui/popup.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import createMetaMaskProvider from "@dimensiondev/metamask-extension-provider";
+import createMetaMaskProvider from "metamask-extension-provider";
 import ReactDOM from "react-dom/client";
 import { browser } from "webextension-polyfill-ts";
 import { HashRouter } from "react-router-dom";
